+++ cli/azd/extensions/azure.coding-agent/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 0.5.1
+
+### Bugs fixed
+
+- Updated message, and help message, when prompting for the Azure subscription to be more descriptive.
+
+## 0.5.0
+
+### Bugs fixed
+
+- Browser now launches properly on Windows
+
 ## 0.4.0
 
 ### Features Added