+++ cli/azd/extensions/azure.foundry.ai.agents/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release History
 
+## 0.0.2 (2025-10-22)
+
+- Support for container app agents from event handlers
+
 ## 0.0.1 (2025-10-09)
 
 - Initial release